WIP: Proof of concept of entity previewer
This properly gets a request through javascript and then is able
to create an updated version of an entity with those values.

It only fails on validation of subfields. When a subfield is
required but its parent is optional then the subfield needn't be
filled if the parent field isn't filled. However in this case an
error is still triggered.