Put widget logic inside of the widget
This is quite a big commit that does a few things. It puts logic
for generating the components of the SEO Analysis widget inside of
the widget iself, instead of using alter hooks for this.

This commit also does away with a lot of configuration that is no
longer needed in the new setup. This might lose some flexibility
that wasn't utilised anyway. It has the advantage of making the
codebase a lot smaller which allows flexibility to be added
wherever needed.

Finally it changes the purpose of the FieldManager and
YoastSEOManager classes somewhat. The FieldManager class will only
be used to manage the attachment of the SEO Analysis field to
entities. It will not be responsible for rendering those fields.
That logic is pushed into the widgets themselves.

The YoastSEOManager will be turned into a more generic service to
provide information about which entities are compatible with the
Real-Time SEO module. It might also provide information about the
status of the RealTime SEO module and should probably be renamed.