Issue #2941887 by Kingdutch: yoast_seo_metatags_alter relies on removed code
The new yoast_seo_metatags_alter searches for a yoast_seo field and
uses its value, if present, to alter the metatags of a page.
Because the field widget settings can not reliably be obbtained
(a widget without those settings may produce the same value) we
instead only check if the value is set. In the widget we then ensure
that no value is present if editing of title or description is
disabled.

This leads to a situtation where a custom title cna be lost if the
field widget setting is disabled and the entity is re-saved,
however, it ensures that the saved value is always equivalent to
the one displayed in the entity create/edit form.